deploy/requirements.txt
constructs>=10.0.0,<11.0.0
aws-cdk-lib
aws-cdk.aws-lambda-python-alpha

